Move SearchResults component to its own subfolder under /components
